backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
pandas>=2.2.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
anthropic>=0.18.0
numpy>=1.24.0
python-dotenv>=1.0.0
